Release nxtvepg-0.7.1 (2002-May-05)
Changes between 0.7.0 and 0.7.1 (as per file CHANGES):
- Windows: exchanged the driver with the free (GPL'ed) DSdrv, a driver
  developed in the DScaler project (see http://dscaler.org).  This
  means that I finally can include a Bt8x8 driver with nxtvepg!  Many thanks
  to e-nek for doing the initial port; also many thanks to John Adcock
  and the other members of the DScaler team.
- Windows: simplified the initial setup procedure by adding the possibility
  to load configuration files or registry of some well known TV applications
  (MoreTV, FreeTV, DScaler, K!TV, MultiDec) to extract their TV card setup.
  The channel tables can also be extracted, to allow for network name
  synchronization (via the Network name configuration dialog)
- Windows: Prepared the possibility to interact with TV applications.  This
  allows for the same features that already were available on UNIX with xawtv
  This feature is disabled in the regular release as there are no TV apps
  yet that support it officially. (The first one will be K!TV, see
  <URL:http://perso.wanadoo.fr/quenottetv/>.  If you are an author of a
  TV app, feel free to mail me if you want to add EPG support to your
  application. A demo application and reference implementation are available.)
- Windows: brought back a lost workaround (lost due to a crash of an almost
  new Deskstar drive - thank you, IBM) in the Tk83.dll for message boxes
  (do not use the native Windows message box from inside Tcl/Tk procedures,
  because it sometimes destroys the main window for unkown reasons)

- Fixed bug in support for big-endian CPU architectures: there was an
  alignment error in one of the database structures which caused an abort
  with a bus error (e.g. on Sun Sparc CPUs); also endian mismatches were
  not reported in acquisition daemon connect attempts.
- Fixed bug in the EPG scan that caused a crash 64-bit architectures
  (e.g. PowerPC). Thanks to Helge Kreutzmann for reporting this bug.
- Added automatic endian conversion during database reload and acquisition
  daemon connections. In earlier versions the reload was simply refused.
  Note: the data is still written in the native format and only converted if
  neccessary, so there is no performance impact for little-endian machines
  (e.g. in contrast to writing always in network byte order)
- Improved channel change handling (handshake between acq master & slave)
- Reduced latency in VPS/PDC decoding after channel changes (the latency
  was increased due to a change in 0.7.0)
- UNIX: improved startup of acquisition process/thread; fixed the design bug
  due to which the TV tuner was programmed even if the VBI could not be opened.
- Improved hyperlinks in help pages: can now link into sub-sections too,
  which is esp. useful for the "Configuration" chapter.
- Fixed a bug in the decoder for language and sub-title information blocks
  (currently only broadcasted in France by M6 and Canal+).

See file CHANGES.RCS for a detailed log of revisions of all files.

+++ epgdb/epgdbif.c
@@ -24,7 +24,7 @@
  *
  *  Author: Tom Zoerner
  *
- *  $Id: epgdbif.c,v 1.40 2002/02/13 21:00:20 tom Exp tom $
+ *  $Id: epgdbif.c,v 1.41 2002/03/29 17:27:08 tom Exp tom $
  */
 
 #define DEBUG_SWITCH DEBUG_SWITCH_EPGDB
@@ -766,7 +766,7 @@ bool EpgDbGetStat( CPDBC dbc, EPGDB_BLOCK_COUNT * pCount, time_t * acqMinTime, u
    const AI_NETWOP *pNetwops;
    uint blockCount[2][MAX_NETWOP_COUNT];
    register uint cur_stream;
-   ulong  count1, count2;
+   uint32_t  count1, count2;
    uchar  ai_version[2];
    double avgPercentage1, avgPercentage2, variance1, variance2;
    uint   acqRepSum[2];
@@ -880,7 +880,7 @@ bool EpgDbGetStat( CPDBC dbc, EPGDB_BLOCK_COUNT * pCount, time_t * acqMinTime, u
          if (count1 > 0)
          {
             pCount[0].ai += count1;
-            blockCount[0][netwop] = (uint)((ulong)1000L * blockCount[0][netwop] / count1);
+            blockCount[0][netwop] = (uint)(1000L * blockCount[0][netwop] / count1);
          }
          else
             blockCount[0][netwop] = 1000L;
@@ -889,7 +889,7 @@ bool EpgDbGetStat( CPDBC dbc, EPGDB_BLOCK_COUNT * pCount, time_t * acqMinTime, u
          if (count2 > 0)
          {
             pCount[1].ai += count2;
-            blockCount[1][netwop] = (uint)((ulong)1000L * blockCount[1][netwop] / count2);
+            blockCount[1][netwop] = (uint)(1000L * blockCount[1][netwop] / count2);
          }
          else
             blockCount[1][netwop] = 1000L;
@@ -934,9 +934,9 @@ bool EpgDbGetStat( CPDBC dbc, EPGDB_BLOCK_COUNT * pCount, time_t * acqMinTime, u
 // - modulo 0x10000
 // - if stop == start - 1  =>  distance == 0  (see ETS 300 707)
 //
-ulong EpgDbGetPiBlockCount( uint startNo, uint stopNo )
+uint EpgDbGetPiBlockCount( uint startNo, uint stopNo )
 {
-   ulong result;
+   uint result;
 
    if (stopNo >= startNo)
    {
